requirements.txt
asgiref==3.5.2; python_version >= "3.8"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.5.0"
certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.5.0"
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
django-cors-headers==3.13.0; python_version >= "3.7"
django-filter==22.1; python_version >= "3.7"
django-location-field==2.1.0
django-phonenumber-field==6.3.0; python_version >= "3.7"
django==4.0.5; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
drf-yasg==1.20.0; python_version >= "3.6"
idna==3.3; python_version >= "3.7" and python_version < "4"
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_version >= "3.6"
jinja2==3.1.2; python_version >= "3.7"
jsonschema==4.6.0; python_version >= "3.7"
markdown==3.3.7; python_version >= "3.6"
markupsafe==2.1.1; python_version >= "3.7"
packaging==21.3; python_version >= "3.6"
phonenumbers==8.12.50; python_version >= "3.7"
pillow==9.1.1; python_version >= "3.7"
pyjwt==2.4.0; python_version >= "3.7"
pyparsing==3.0.9; python_full_version >= "3.6.8" and python_version >= "3.6"
pyrsistent==0.18.1; python_version >= "3.7"
python-dotenv==0.20.0; python_version >= "3.5"
pytz==2022.1; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.6"
requests==2.28.0; python_version >= "3.7" and python_version < "4"
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.6"
ruamel.yaml==0.17.21; python_version >= "3.6"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.3.0"
sqlparse==0.4.2; python_version >= "3.8"
swagger-spec-validator==2.7.4; python_version >= "3.6"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.7"
